Streamlining and added custom separator.
Rewrote and streamlined the command line parsing and how
options are propagated to the file parser.
Command line options can now be placed anywhere on the
command line.
Added the ability to change the column separator in the
output file via the the -u flag (default is a single space).